+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `toolbar top` and `toolbar bottom` components (Orange-OpenSource/ouds-ios#1174)
 - `text area` component (Orange-OpenSource/ouds-ios#543)
 - `PIN code input` component (Orange-OpenSource/ouds-ios#998)
 - Agentic AI dedicated files (Orange-OpenSource/ouds-ios#1399)